agno_cli/pyproject.toml
[project]
name = "agno-cli"
version = "0.1.0"
description = "CLI for interacting with AgentOS instances"
requires-python = ">=3.9,<4"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  {name = "Agno", email = "support@agno.com"}
]
keywords = [
  "agent",
  "cli",
  "agentos",
]
classifiers = [
    "Intended Audience :: Developers",
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
]

dependencies = [
  "httpx",
  "pydantic",
  "pyyaml",
  "typer",
]

[project.optional-dependencies]
dev = [
  "mypy==1.18.2",
  "ruff==0.14.3",
  "pytest",
]

[project.scripts]
agno-os = "agno_cli.main:app"

[project.urls]
homepage = "https://agno.com"
documentation = "https://docs.agno.com"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agno_cli*"]

[tool.setuptools.package-data]
agno_cli = ["py.typed"]

[tool.ruff]
line-length = 120
target-version = "py39"
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
